refactor(mapper): 将用户名字段改为昵称字段
- 在 ProjectMapper.xml 中将 su.user_name 改为 su.nick_name 作为 createName
- 在 ProjectMapper.xml 中将 user.user_name 改为 user.nick_name 作为 principalName- 在 SlideMapper.xml 中将 su.user_name 改为 su.nick_name 作为 user_name

+++ src/main/java/cn/staitech/fr/service/impl/ImageServiceImpl.java
@@ -134,6 +134,9 @@ public class ImageServiceImpl extends ServiceImpl<ImageMapper, Image>
                     .eq(Slide::getDelFlag, DEL_FLAG_NORMAL)
                     .select(Slide::getImageId));
             forbidIds = slideList.stream().map(Slide::getImageId).collect(Collectors.toList());
+            if (CollectionUtils.isNotEmpty(forbidIds)){
+                throw new RuntimeException("已关联项目的图像不可删除，请重新选择");
+            }
             imageIdList.removeAll(forbidIds);
             if (CollectionUtils.isNotEmpty(imageIdList)){
                 imageMapper.deleteBatchIds(imageIdList);
